scripts/requirements.txt
selenium>=4.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
numpy>=1.24.0
webdriver-manager>=4.0.0
